% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLUBmodels.R
\name{print.compareLUBmodels}
\alias{print.compareLUBmodels}
\title{Print method for LUB model comparisons}
\usage{
\method{print}{compareLUBmodels}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"compareLUBmodels"}.}

\item{digits}{Number of digits used for printing.}

\item{...}{Additional arguments.}
}
\value{
The input object, returned invisibly.
}
\description{
Print method for LUB model comparisons
}
