% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLUB_mle.R
\name{profileLUB}
\alias{profileLUB}
\title{Profile Likelihood for the LUB Parameter}
\usage{
profileLUB(
  x,
  size,
  beta.grid = exp(seq(log(0.05), log(30), length.out = 300L)),
  tol = 1e-14,
  maxit = 10000L
)
}
\arguments{
\item{x}{Observed bounded counts.}

\item{size}{Positive integer giving the upper endpoint of the support.}

\item{beta.grid}{Positive numeric grid of beta values.}

\item{tol}{Numerical tolerance passed to `dLUB()`.}

\item{maxit}{Maximum number of hypergeometric-series terms.}
}
\value{
An object of class `"profileLUB"`.
}
\description{
Evaluates the log-likelihood of the Lambert-uniform binomial model
over a user-specified grid of positive beta values.
}
\examples{
x <- c(0, 0, 1, 0, 2, 1, 0, 0)
pr <- profileLUB(x, size = 4)
pr

}
