% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLUBmodels.R
\name{summary.compareLUBmodels}
\alias{summary.compareLUBmodels}
\title{Summary method for LUB model comparisons}
\usage{
\method{summary}{compareLUBmodels}(object, digits = 4, ...)
}
\arguments{
\item{object}{An object of class \code{"compareLUBmodels"}.}

\item{digits}{Number of digits used for printing.}

\item{...}{Additional arguments.}
}
\value{
The input object, returned invisibly.
}
\description{
Summary method for LUB model comparisons
}
